Crate hotshot_constants

configurable constants for hotshot

Structs

Version

Type for protocol version number

Constants

COMBINED_NETWORK_CACHE_SIZE

the number of messages to cache in the combined network

COMBINED_NETWORK_MIN_PRIMARY_FAILURES

the number of messages to attempt to send over the primary network before switching to prefer the secondary network

COMBINED_NETWORK_PRIMARY_CHECK_INTERVAL

the number of messages to send over the secondary network before re-attempting the (presumed down) primary network

KAD_DEFAULT_REPUB_INTERVAL_SEC

the default kademlia record republication interval (in seconds)

LOOK_AHEAD

the number of views to gather information for ahead of time

VERSION_0_1

Constant for protocol version 0.1.